Added redirection example
